refactor authentication package
