+++ lib/src/features/authentication/data/remote/dtos/register_dto.dart
@@ -1,6 +1,5 @@
 import 'package:equatable/equatable.dart';
 import 'package:json_annotation/json_annotation.dart';
-import 'package:vinhcine/src/features/authentication/domain/data/remote/dtos/payload_dto.dart';
 
 part 'register_dto.g.dart';
