+++ lib/src/features/authentication/data/remote/services/auth_service.dart
@@ -1,6 +1,7 @@
 import 'package:dio/dio.dart';
 import 'package:retrofit/retrofit.dart';
-import 'package:vinhcine/src/features/authentication/domain/data/remote/dtos/get_token_response.dart';
+
+import '../dtos/get_token_response.dart';
 
 part 'auth_service.g.dart';
