+++ lib/src/features/authentication/domain/model/register_mapper.dart
@@ -1,4 +1,4 @@
-import 'package:vinhcine/src/features/authentication/domain/data/remote/dtos/register_response.dart';
+import 'package:vinhcine/src/features/authentication/data/remote/dtos/register_response.dart';
 import 'package:vinhcine/src/features/authentication/domain/model/register.dart';
 
 extension RegisterData on RegisterResponse {
