+++ lib/src/features/authentication/domain/model/signin_mapper.dart
@@ -1,4 +1,5 @@
-import 'package:vinhcine/src/features/authentication/domain/data/remote/dtos/get_token_response.dart';
+
+import 'package:vinhcine/src/features/authentication/data/remote/dtos/get_token_response.dart';
 
 extension SignInData on GetTokenResponse {
   String toModel() => token;
