+++ lib/src/features/authentication/domain/repositories/register_repository_mapper.dart
@@ -1,5 +1,5 @@
 
-import 'package:vinhcine/src/features/authentication/domain/data/remote/dtos/register_response.dart';
+import 'package:vinhcine/src/features/authentication/data/remote/dtos/register_response.dart';
 import 'package:vinhcine/src/features/authentication/domain/model/register.dart';
 import 'package:vinhcine/src/features/authentication/domain/model/register_mapper.dart';
