+++ lib/src/features/authentication/domain/repositories/signin_repository_mapper.dart
@@ -1,4 +1,5 @@
-import '../data/remote/dtos/get_token_response.dart';
+
+import '../../data/remote/dtos/get_token_response.dart';
 
 class SignInMapper {
   static String mapToModel(GetTokenResponse response) {
